:truck: recycle: Move payment models into models folder, add comments to code files
