--- libs/private/functions/payments/core/src/lib/private-functions-payments-core.ts
@@ -1,3 +1,0 @@
-export function privateFunctionsPaymentsCore(): string {
-  return 'private-functions-payments-core';
-}
