+++ libs/private/model/payments-models/jest.config.ts
