+++ libs/private/model/payments-models/src/lib/payment-method.ts
@@ -1,17 +1,14 @@
+/**
+ * various payments methods available, working with cards for now.
+ * metadata is typed as any, will be determined as project moves on
+ */
 export interface PaymentMethod {
     paymentType:    PaymentType
     metadata:       any;
 }
-
 
 export enum PaymentType {
     debitCard = 1,
     creditCard  = 2,
     mPesa = 3
 }
-
-/**comments  on type and params*/
-
-//end point api/rcv payment 
-//handler
-//private/model/payments
