+++ libs/private/model/payments-models/src/lib/sequence-type.ts
@@ -1,0 +1,8 @@
+/**
+ * Duration in which a subscription is renewed. 
+ */
+export enum SequenceType {
+    Recurring = 1,
+    BiAnnual = 2,
+    Quarterly = 3
+}
