+++ libs/private/model/payments-models/src/lib/subscription.ts
@@ -1,0 +1,14 @@
+import { Amount } from "./payment";
+
+/**
+ * Interval takes types monthy, weekly, yearly
+ * @see https://docs.mollie.com/payments/recurring on how to handle recurring payments
+ */
+export interface Subscription {
+    orgId:         string;
+    amount:        Amount;
+    interval:      string;
+    times?:        number
+    description:   string;
+}
+
