--- libs/private/payments-models/src/lib/customer.ts
@@ -1,8 +1,0 @@
-
-/**The customer id is returned by the customer API and will be saved to db */
-export interface Customer{
-    customerId:     string
-    name:           string;
-    email:          string;
-    isPremium:      boolean;
-}
