--- libs/private/payments-models/src/lib/sequence-type.ts
@@ -1,5 +1,0 @@
-export enum SequenceType {
-    recurring = 1,
-    biAnnual = 2,
-    quarterly = 3
-}
