--- libs/private/payments-models/src/lib/subscription.ts
@@ -1,10 +1,0 @@
-import { Amount } from "./payment";
-import { PaymentMethod } from "./payment-method";
-
-export interface Subscription {
-    amount: Amount
-    interval: string;
-    times?: number
-    description: string;
-    method: PaymentMethod;
-}
