--- libs/private/payments-models/src/lib/transaction.ts
@@ -1,5 +1,0 @@
-export interface Transaction {
-    amount: string;
-    date: string;
-    tokens?: string;
-}
